Move League Of Legends to Games
