+++ PyChan/Core/Commands/Games/games.py
@@ -1,7 +1,8 @@
-import nextcord
 from nextcord.ext import commands
-from Core.Commands.Games.Functions.osu import Osu
 
+from Core.Commands.Games.Functions.osu import Osu
+from Core.Commands.Games.Functions.summoner_info import SummonerInfo
+from Core.Commands.Games.Functions.lastMatch import LastMatch
 
 class Games(commands.Cog):
     """Class contains gaming methods
@@ -12,3 +13,5 @@ class Games(commands.Cog):
         """
         self.bot = bot
         self.bot.add_cog(Osu(bot))
+        self.bot.add_cog(SummonerInfo(bot))
+        self.bot.add_cog(LastMatch(bot))
