+++ PyChan/Core/Commands/commands.py
@@ -1,9 +1,7 @@
-import nextcord
 from nextcord.ext import commands
 
 from Core.Commands.Help.help import Help
 from Core.Commands.Images.images import Images
-from Core.Commands.Others.others import Others
 from Core.Commands.Science.science import Science
 from Core.Commands.Settings.settings import Settings
 from Core.Commands.SKNIKOD.skni_kod import SKNIKOD
@@ -22,7 +20,6 @@ class Commands(commands.Cog):
         self.bot = bot
         self.bot.add_cog(Help(bot))
         self.bot.add_cog(Images(bot))
-        self.bot.add_cog(Others(bot))
         self.bot.add_cog(Science(bot))
         self.bot.add_cog(SKNIKOD(bot))
         self.bot.add_cog(Settings(bot))
